feat: Refactor FlowBuilder component to enhance node and edge management
- Introduced custom node and edge types for better structure and clarity.
- Updated initial nodes and edges to reflect a coffee ordering flow.
- Implemented context menu for adding and deleting nodes.
- Added auto-arrangement of nodes on component mount.
- Enhanced connection validation logic to prevent invalid connections.
- Integrated node editing dialogs for improved user interaction.
- Removed deprecated StepNode and ToolNode components.
- Updated UI components for badges and buttons to use forward refs for better accessibility.
