+++ tools/visual-builder/src/components/nodes/StepNode.tsx
@@ -1,80 +1,97 @@
 import { memo } from 'react';
 import { Handle, Position, type NodeProps } from '@xyflow/react';
 import { Button } from '../ui/button';
-import { Edit3, Settings } from 'lucide-react';
-import type { StepNodeData } from '../../models/nomos';
+import { Edit, Play } from 'lucide-react';
+import { useFlowContext } from '../../context/FlowContext';
+import type { StepNodeData } from '../../types';
 
-export const StepNode = memo(({ data, selected }: NodeProps<StepNodeData>) => {
+export const StepNode = memo((props: NodeProps) => {
+  const { setEditingNode } = useFlowContext();
+  const data = props.data as StepNodeData;
+
+  const handleEdit = () => {
+    setEditingNode(props.id, 'step', data);
+  };
+
   return (
-    <div
-      className={`
-        relative bg-white border-2 rounded-lg shadow-sm min-w-[200px] p-4
-        ${selected ? 'border-gray-800' : 'border-gray-300'}
-        hover:border-gray-600 transition-colors
-      `}
-    >
-      {/* Input Handle */}
-      <Handle
-        type="target"
-        position={Position.Top}
-        className="w-3 h-3 !bg-gray-600 !border-2 !border-white"
-      />
-
-      {/* Node Header */}
-      <div className="flex items-center justify-between mb-3">
-        <div className="flex items-center gap-2">
-          <Settings className="w-4 h-4 text-gray-600" />
-          <span className="font-medium text-gray-900">Step</span>
+    <div className="bg-white border border-gray-300 rounded shadow-sm w-[280px] hover:border-gray-400 transition-colors">
+      {/* Header */}
+      <div className="bg-gray-50 px-3 py-2 border-b border-gray-200 rounded-t flex items-center justify-between">
+        <div className="flex items-center gap-2 min-w-0 flex-1">
+          <Play className="w-4 h-4 text-gray-600 flex-shrink-0" />
+          <span className="font-medium text-sm text-gray-700 truncate">{data.step_id}</span>
         </div>
-        <Button
-          variant="ghost"
-          size="sm"
-          className="w-6 h-6 p-0 hover:bg-gray-100"
-        >
-          <Edit3 className="w-3 h-3" />
+        <Button variant="ghost" size="sm" className="h-6 w-6 p-0 hover:bg-gray-200 flex-shrink-0" onClick={handleEdit}>
+          <Edit className="w-3 h-3" />
         </Button>
       </div>
 
-      {/* Step ID */}
-      <div className="mb-2">
-        <div className="text-xs text-gray-500 mb-1">Step ID</div>
-        <div className="text-sm font-mono bg-gray-50 px-2 py-1 rounded border">
-          {data.step_id}
-        </div>
+      {/* Body */}
+      <div className="p-3 space-y-2">
+        {data.description && (
+          <div className="text-sm text-gray-600 line-clamp-2">
+            {data.description}
+          </div>
+        )}
+
+        {data.available_tools && data.available_tools.length > 0 && (
+          <div className="flex flex-wrap gap-1 mt-2">
+            {data.available_tools.slice(0, 3).map((tool) => (
+              <span
+                key={tool}
+                className="px-2 py-1 bg-blue-100 text-blue-700 text-xs rounded-full truncate max-w-[80px]"
+                title={tool}
+              >
+                {tool}
+              </span>
+            ))}
+            {data.available_tools.length > 3 && (
+              <span className="px-2 py-1 bg-gray-100 text-gray-600 text-xs rounded-full">
+                +{data.available_tools.length - 3}
+              </span>
+            )}
+          </div>
+        )}
+
+        {data.routes && data.routes.length > 0 && (
+          <div className="text-xs text-gray-500 truncate">
+            Routes to: {data.routes.map(r => r.target).join(', ')}
+          </div>
+        )}
+
+        {data.auto_flow && (
+          <div className="text-xs text-green-600 font-medium">
+            Auto Flow
+          </div>
+        )}
       </div>
 
-      {/* Persona (if exists) */}
-      {data.persona && (
-        <div className="mb-2">
-          <div className="text-xs text-gray-500 mb-1">Persona</div>
-          <div className="text-xs text-gray-700 line-clamp-2">
-            {data.persona}
-          </div>
-        </div>
-      )}
-
-      {/* Tools Count */}
-      {data.tools && data.tools.length > 0 && (
-        <div className="mb-2">
-          <div className="text-xs text-gray-500">
-            Tools: {data.tools.length}
-          </div>
-        </div>
-      )}
-
-      {/* Routes Count */}
-      {data.routes && Object.keys(data.routes).length > 0 && (
-        <div className="text-xs text-gray-500">
-          Routes: {Object.keys(data.routes).length}
-        </div>
-      )}
-
-      {/* Output Handle */}
+      {/* Handles */}
+      {/* Top handle for incoming step connections */}
+      <Handle
+        type="target"
+        position={Position.Top}
+        id="step-input"
+        className="w-3 h-3 !bg-gray-600 border-2 border-white"
+      />
+      
+      {/* Bottom handle for outgoing step connections */}
       <Handle
         type="source"
         position={Position.Bottom}
-        className="w-3 h-3 !bg-gray-600 !border-2 !border-white"
+        id="step-output"
+        className="w-3 h-3 !bg-gray-600 border-2 border-white"
+      />
+      
+      {/* Right handle for tool connections */}
+      <Handle
+        type="source"
+        position={Position.Right}
+        id="tool-output"
+        className="w-2 h-2 !bg-blue-500 border-2 border-white"
       />
     </div>
   );
 });
+
+StepNode.displayName = 'StepNode';
