+++ tools/visual-builder/src/components/nodes/ToolNode.tsx
@@ -1,74 +1,57 @@
 import { memo } from 'react';
 import { Handle, Position, type NodeProps } from '@xyflow/react';
 import { Button } from '../ui/button';
-import { Edit3, Tool } from 'lucide-react';
-import type { ToolNodeData } from '../../models/nomos';
+import { Edit, Wrench } from 'lucide-react';
+import { useFlowContext } from '../../context/FlowContext';
+import type { ToolNodeData } from '../../types';
 
-export const ToolNode = memo(({ data, selected }: NodeProps<ToolNodeData>) => {
+export const ToolNode = memo((props: NodeProps) => {
+  const { setEditingNode } = useFlowContext();
+  const data = props.data as ToolNodeData;
+
+  const handleEdit = () => {
+    setEditingNode(props.id, 'tool', data);
+  };
+
+  const paramCount = data.parameters ? Object.keys(data.parameters).length : 0;
+
   return (
-    <div
-      className={`
-        relative bg-white border-2 rounded-lg shadow-sm min-w-[180px] p-4
-        ${selected ? 'border-gray-800' : 'border-gray-300'}
-        hover:border-gray-600 transition-colors
-      `}
-    >
-      {/* Input Handle */}
-      <Handle
-        type="target"
-        position={Position.Left}
-        className="w-3 h-3 !bg-gray-600 !border-2 !border-white"
-      />
-
-      {/* Node Header */}
-      <div className="flex items-center justify-between mb-3">
-        <div className="flex items-center gap-2">
-          <Tool className="w-4 h-4 text-gray-600" />
-          <span className="font-medium text-gray-900">Tool</span>
+    <div className="bg-blue-50 border border-blue-300 rounded shadow-sm w-[200px] hover:border-blue-400 transition-colors">
+      {/* Header */}
+      <div className="bg-blue-100 px-3 py-2 border-b border-blue-200 rounded-t flex items-center justify-between">
+        <div className="flex items-center gap-2 min-w-0 flex-1">
+          <Wrench className="w-4 h-4 text-blue-600 flex-shrink-0" />
+          <span className="font-medium text-sm text-blue-800 truncate">{data.name}</span>
         </div>
-        <Button
-          variant="ghost"
-          size="sm"
-          className="w-6 h-6 p-0 hover:bg-gray-100"
-        >
-          <Edit3 className="w-3 h-3" />
+        <Button variant="ghost" size="sm" className="h-6 w-6 p-0 hover:bg-blue-200 flex-shrink-0" onClick={handleEdit}>
+          <Edit className="w-3 h-3" />
         </Button>
       </div>
 
-      {/* Tool ID */}
-      <div className="mb-2">
-        <div className="text-xs text-gray-500 mb-1">Tool ID</div>
-        <div className="text-sm font-mono bg-gray-50 px-2 py-1 rounded border">
-          {data.tool_id}
-        </div>
-      </div>
-
-      {/* Description (if exists) */}
-      {data.description && (
-        <div className="mb-2">
-          <div className="text-xs text-gray-500 mb-1">Description</div>
-          <div className="text-xs text-gray-700 line-clamp-2">
+      {/* Body */}
+      <div className="p-3 space-y-2">
+        {data.description && (
+          <div className="text-sm text-blue-700 line-clamp-2">
             {data.description}
           </div>
-        </div>
-      )}
+        )}
 
-      {/* Package Reference (if exists) */}
-      {data.package_reference && (
-        <div className="mb-2">
-          <div className="text-xs text-gray-500 mb-1">Package</div>
-          <div className="text-xs font-mono bg-gray-50 px-2 py-1 rounded border">
-            {data.package_reference}
+        {paramCount > 0 && (
+          <div className="text-xs text-blue-600">
+            {paramCount} parameter{paramCount !== 1 ? 's' : ''}
           </div>
-        </div>
-      )}
+        )}
+      </div>
 
-      {/* Parameters Count */}
-      {data.parameters && Object.keys(data.parameters).length > 0 && (
-        <div className="text-xs text-gray-500">
-          Parameters: {Object.keys(data.parameters).length}
-        </div>
-      )}
+      {/* Handles */}
+      <Handle
+        type="target"
+        position={Position.Left}
+        id="tool-input"
+        className="w-2 h-2 !bg-blue-500 border-2 border-white"
+      />
     </div>
   );
 });
+
+ToolNode.displayName = 'ToolNode';
